fix(product): move workflowRunStatusDotTone out of domain, gate was red
check_frontend_boundaries.py flagged domain/workflows/run-presentation.ts
importing StatusDotTone from #product/primitives/StatusDot -- a UI
primitive import in a ProductClient domain module, which the boundary
gate limits to Cloud SDK types, AnyHarness contract types/pure helpers, and
test-only vitest. This mapping was added in the earlier workflows-area
conformance commit in this same slice, so this is a follow-up fix rather
than a new violation from WP-4.

Moved the tone-to-StatusDotTone map and workflowRunStatusDotTone() into a
new components/workflows/workflow-run-status-dot.ts, mirroring the pattern
SubagentRosterRow.tsx and TerminalRosterRow.tsx already use for their own
local tone maps -- the mapping lives in the UI layer that owns StatusDot,
not in domain. run-presentation.ts keeps WorkflowRunTone as a plain string
union with no UI dependency. WorkflowRunDetail.tsx and WorkflowRunList.tsx
now import workflowRunStatusDotTone from the new local file instead of
domain.

check_frontend_boundaries.py passes. Re-ran the workflows-area tests that
exercise both consumers: WorkflowRuns.test.tsx (4/4, covers
WorkflowRunDetail) and runs/WorkflowRunsSurface.test.tsx (14/14, covers
WorkflowRunList).
