+++ apps/packages/product-client/src/components/workflows/WorkflowRunDetail.tsx
@@ -5,7 +5,7 @@ import type {
   WorkflowRunPresentation,
   WorkflowRunTone,
 } from "#product/domain/workflows/run-presentation";
-import { workflowRunStatusDotTone } from "#product/domain/workflows/run-presentation";
+import { workflowRunStatusDotTone } from "#product/components/workflows/workflow-run-status-dot";
 import { Button } from "#product/primitives/Button";
 import { StatusDot } from "#product/primitives/StatusDot";
 import { Card } from "#product/primitives/patterns/Card";
